Condense two separate map classes into one
